Fix register type - need to be int32

+++ src/GNR1_Firmware/rev_E/NRF/MD_MASTER/FTEX_MODULES/Inc/vc_interface.h
@@ -210,7 +210,7 @@ typedef enum
 
 int32_t VCI_ReadRegister(VCI_Handle_t* pHandle, uint16_t RegID);
 
-void VCI_SetRegister(VCI_Handle_t* pHandle, uint16_t RegID, uint8_t value);
+void VCI_SetRegister(VCI_Handle_t* pHandle, uint16_t RegID, int32_t value);
